refactor: переименовывает компоненты event в point
