+++ src/components/trip/points/collapsed-points.js
