+++ src/components/trip/points/editable-point.js
