+++ src/components/trip/points/points-list.js
